+++ code/Core/System/Win32/ProcessWin32.cpp
@@ -35,22 +35,22 @@ public:
 		return false;
 	}
 
-	virtual int tell() const
+	virtual int64_t tell() const
 	{
 		return 0;
 	}
 
-	virtual int available() const
+	virtual int64_t available() const
 	{
 		return 0;
 	}
 
-	virtual int seek(SeekOriginType origin, int offset)
+	virtual int64_t seek(SeekOriginType origin, int64_t offset)
 	{
 		return 0;
 	}
 
-	virtual int read(void* block, int nbytes)
+	virtual int64_t read(void* block, int64_t nbytes)
 	{
 		bool processTerminated = (WaitForSingleObject(m_hProcess, 0) == WAIT_OBJECT_0);
 
@@ -86,10 +86,10 @@ public:
 				return -1;
 		}
 
-		return int(dwRead);
+		return int64_t(dwRead);
 	}
 
-	virtual int write(const void* block, int nbytes)
+	virtual int64_t write(const void* block, int64_t nbytes)
 	{
 		return 0;
 	}
